use dynamic component
